requirements.txt
requests
python-dotenv
TTS
torch